WIP - first pass at adding scheme scripting. Functional but slow.
Also, use of scheme vs. a rule file is hardcoded at the moment.
Needs a way to define scheme file to load in data file, or rule file to
load.
Might need to pass the full board to scheme, or register neighbor
counting as a native function for speed.
Might also need to use some forms of cacheing